Refactors transfer retrieval to use nonce and improve flow
Streamlines transfer data fetching by introducing a nonce retrieval step and reworking the retrieval logic to iterate over all subjects for an institution.
Prepares for more robust pagination and data handling in future updates.

Note the same course can be on multiple pages
Additionally, We can skip the 1st page, as we get that when we fetch the total pages
Will continue working on this post work